refactor: extract FIM prompt building to dedicated fim.ts file
- Create new FimPromptBuilder class in fim.ts
- Move getFimPrompts method from HoleFiller to FimPromptBuilder
- Update GhostInlineCompletionProvider to use FimPromptBuilder
- Separates FIM-specific logic from chat-based completion logic
- All tests pass (63/63)
