+++ src/services/ghost/classic-auto-complete/HoleFiller.ts
@@ -1,7 +1,6 @@
 import { AutocompleteInput } from "../types"
 import { GhostContextProvider } from "./GhostContextProvider"
 import { formatSnippets } from "../../continuedev/core/autocomplete/templating/formatting"
-import { getTemplateForModel } from "../../continuedev/core/autocomplete/templating/AutocompleteTemplate"
 
 export interface FillInAtCursorSuggestion {
 	text: string
@@ -170,40 +169,5 @@ ${formattedContext}${formattedContext ? "\n" : ""}${helper.prunedPrefix}{{FILL_H
 TASK: Fill the {{FILL_HERE}} hole. Answer only with the CORRECT completion, and NOTHING ELSE. Do it now.
 Return the COMPLETION tags`
 		)
-	}
-
-	/**
-	 * Build FIM (Fill-In-the-Middle) prompts with formatted prefix and pruned suffix
-	 */
-	async getFimPrompts(
-		autocompleteInput: AutocompleteInput,
-		modelName: string,
-	): Promise<{
-		formattedPrefix: string
-		prunedSuffix: string
-	}> {
-		const { filepathUri, helper, snippetsWithUris, workspaceDirs } =
-			await this.contextProvider.getProcessedSnippets(autocompleteInput, autocompleteInput.filepath)
-
-		// Use pruned prefix/suffix from HelperVars (token-limited based on DEFAULT_AUTOCOMPLETE_OPTS)
-		const prunedPrefixRaw = helper.prunedPrefix
-		const prunedSuffix = helper.prunedSuffix
-
-		const template = getTemplateForModel(modelName)
-
-		let formattedPrefix = prunedPrefixRaw
-		if (template.compilePrefixSuffix && prunedSuffix) {
-			const [compiledPrefix] = template.compilePrefixSuffix(
-				prunedPrefixRaw,
-				prunedSuffix,
-				filepathUri,
-				"", // reponame not used in our context
-				snippetsWithUris,
-				workspaceDirs,
-			)
-			formattedPrefix = compiledPrefix
-		}
-
-		return { formattedPrefix, prunedSuffix }
 	}
 }
